Put generateGrammarCommandHandler into separate file as discussed in PR #1093 removed async keyword again also changed some * as imports to destructured imports
